refactor: move unit tests to __tests__ folders
