+++ src/__tests__/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 describe('App', (): void => {
   it('renders the application', (): void => {
