+++ src/components/__tests__/Screen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Screen from './Screen';
+import Screen from '../Screen';
 
 describe('Screen', (): void => {
   it('renders the screen component', (): void => {
