+++ src/components/__tests__/SlimSpinArrow.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import SlimSpinArrow from './SlimSpinArrow';
+import SlimSpinArrow from '../SlimSpinArrow';
 
 // TODO: Add React.memo() related unit tests
 describe('SlimSpinArrow', (): void => {
